Move RgbPixel to a separate module
